aabot/requirements.txt
discord.py
httpx
fuzzywuzzy[speedup]
dacite
table2ascii
pytz
pynacl
html2text
pydantic
psycopg[binary]
sqlalchemy
